docs: Add TODO to make the API async
We will have large files and creating the hash may take minutes or
hours. When scaling up, we definitely will want to move this to an
architecture with a job queue of some sort.

Hence we should start off with an async API design, so that we don't
paint ourselves into a corner by offering a blocking/sync API now.